+++ plugins/plotly-express/src/js/src/PlotlyExpressPlugin.ts
@@ -1,10 +1,10 @@
 import { type WidgetPlugin, PluginType } from '@deephaven/plugin';
 import { vsGraph } from '@deephaven/icons';
-import type { Widget } from '@deephaven/jsapi-types';
+import type { dh } from '@deephaven/jsapi-types';
 import { PlotlyExpressChart } from './PlotlyExpressChart.js';
 import { PlotlyExpressChartPanel } from './PlotlyExpressChartPanel.js';
 
-export const PlotlyExpressPlugin: WidgetPlugin<Widget> = {
+export const PlotlyExpressPlugin: WidgetPlugin<dh.Widget> = {
   name: '@deephaven/plotly-express',
   type: PluginType.WIDGET_PLUGIN,
   supportedTypes: 'deephaven.plot.express.DeephavenFigure',