+++ plugins/ui/src/deephaven/ui/components/list_action_group.py
@@ -11,10 +11,12 @@ ActionGroupItem = Union[Stringable, ItemElement]
 ListActionGroupElement = Element
 
 
+# TODO: pydocs for list_action_group #483
 def list_action_group(
     *children: ActionGroupItem,
     on_action: Callable[[ActionKey, Key], None] | None = None,
     on_selection_change: Callable[[Selection, Key], None] | None = None,
+    on_change: Callable[[Selection, Key], None] | None = None,
     **props: Any,
 ) -> ListActionGroupElement:
     """
@@ -26,6 +28,9 @@ def list_action_group(
         on_action: Handler that is called when an item is pressed.
             The first argument is the key of the action, the second argument is the key of the list_view item.
         on_selection_change: Handler that is called when the selection changes.
+            The first argument is the selection, the second argument is the key of the list_view item.
+        on_change: Alias of on_selection_change.
+            Handler that is called when the selection changes.
             The first argument is the selection, the second argument is the key of the list_view item.
         **props: Any other ActionGroup prop.