+++ plugins/ui/src/deephaven/ui/components/list_view.py
@@ -1,61 +1,59 @@
 from __future__ import annotations
 
-from typing import Callable, Any, Union
+from typing import Callable, Any
 
 from deephaven.table import Table
 
-from .item import ItemElement
-from .list_action_group import ListActionGroupElement
-from .list_action_menu import ListActionMenuElement
+from .item_table_source import ItemTableSource
 from ..elements import BaseElement, Element
-from .._internal.utils import create_props
-from ..types import ColumnName, Stringable, Selection
+from .._internal.utils import create_props, unpack_item_table_source
+from .item import Item
+from ..types import Density, Selection, SelectionMode
 
-ListViewItem = Union[Stringable, ItemElement]
 ListViewElement = Element
+
+SUPPORTED_SOURCE_ARGS = {
+    "key_column",
+    "label_column",
+    "description_column",
+    "icon_column",
+    "actions",
+}
 
 
 def list_view(
-    *children: ListViewItem | Table,
-    key_column: ColumnName | None = None,
-    label_column: ColumnName | None = None,
-    description_column: ColumnName | None = None,
-    icon_column: ColumnName | None = None,
-    actions: ListActionGroupElement | ListActionMenuElement | None = None,
+    *children: Item | Table | ItemTableSource,
+    density: Density | None = "COMPACT",
     default_selected_keys: Selection | None = None,
     selected_keys: Selection | None = None,
+    selection_mode: SelectionMode | None = "MULTIPLE",
     render_empty_state: Element | None = None,
     on_selection_change: Callable[[Selection], None] | None = None,
     on_change: Callable[[Selection], None] | None = None,
     **props: Any,
 ) -> ListViewElement:
     """
-    A list view that can be used to create a list of items. Children should be one of two types:
-    1. If children are of type `ListViewItem`, they are the list items.
+    A list view that can be used to create a list of items. Children should be one of three types:
+    1. If children are of type `Item`, they are the list items.
     2. If children are of type `Table`, the values in the table are the list items.
         There can only be one child, the `Table`.
-
+        The first column is used as the key and label by default.
+    3. If children are of type `ItemTableSource`, complex items are created from the source.
+        There can only be one child, the `ItemTableSource`.
+        Supported `ItemTableSource` arguments are `key_column`, `label_column`, `description_column`,
+        `icon_column`, and `actions`.
 
     Args:
         *children: The options to render within the list_view.
-        key_column:
-            Only valid if children are of type Table.
-            The column of values to use as item keys. Defaults to the first column.
-        label_column:
-            Only valid if children are of type Table.
-            The column of values to display as primary text. Defaults to the key_column value.
-        description_column:
-            Only valid if children are of type Table.
-            The column of values to display as descriptions.
-        icon_column: Only valid if children are of type Table.
-            The column of values to map to icons.
-        actions:
-            Only valid if children are of type Table.
-            The action group or menus to render for all elements within the list view.
+        density:
+            Sets the amount of vertical padding within each cell.
         default_selected_keys:
             The initial selected keys in the collection (uncontrolled).
         selected_keys:
             The currently selected keys in the collection (controlled).
+        selection_mode:
+            By default `"MULTIPLE"`, which allows multiple selection.
+            May also be `"SINGLE"` to allow only single selection, or `"None"`/`None` to allow no selection.
         render_empty_state:
             Sets what the `list_view` should render when there is no content to display.
         on_selection_change:
@@ -69,5 +67,7 @@ def list_view(
         The rendered ListView.
     """
     children, props = create_props(locals())
+
+    children, props = unpack_item_table_source(children, props, SUPPORTED_SOURCE_ARGS)
 
     return BaseElement("deephaven.ui.components.ListView", *children, **props)