+++ plugins/ui/src/deephaven/ui/components/section.py
@@ -1,19 +1,15 @@
 from __future__ import annotations
 
-from typing import Union, Any
+from typing import Any
 
 from .._internal.utils import create_props
 from ..elements import Element, BaseElement
-from ..types import Stringable
-from .item import ItemElement
+from .item import Item
 
-PickerItem = Union[Stringable, ItemElement]
 SectionElement = Element
 
 
-def section(
-    *children: PickerItem, title: str | None = None, **props: Any
-) -> SectionElement:
+def section(*children: Item, title: str | None = None, **props: Any) -> SectionElement:
     """
     A section that can be added to a menu, such as a picker. Children are the dropdown options.