+++ plugins/ui/src/deephaven/ui/components/spectrum/__init__.py
@@ -6,7 +6,9 @@ from .checkbox import *
 from .content import *
 from .range_slider import *
 from .slider import *
+from .contextual_help import *
 from .text_field import *
 from .toggle_button import *
 from .flex import *
 from .date_picker import *
+from .combo_box import *