+++ plugins/ui/src/deephaven/ui/components/spectrum/combo_box.py
@@ -1,0 +1,9 @@
+from typing import Literal
+
+MenuTriggerAction = Literal["focus", "input", "manual"]
+Align = Literal["start", "end"]
+Direction = Literal["bottom", "top"]
+LoadingState = Literal[
+    "loading", "sorting", "loadingMore", "error", "idle", "filtering"
+]
+FormValue = Literal["key", "text"]