+++ plugins/ui/src/deephaven/ui/components/spectrum/validate.py
@@ -9,3 +9,5 @@ TextFieldInputMode = Literal[
 ]
 TextFieldValidationState = Literal["valid", "invalid"]
 NecessityIndicator = Literal["icon", "label"]
+
+ContextualHelperVariant = Literal["help", "info"]