+++ plugins/ui/src/js/src/widget/WidgetTypes.ts
@@ -24,3 +24,24 @@ export type ReadonlyWidgetData = Readonly<WidgetData>;
 
 /** Contains an update for widget data. Only the keys that are updated are passed. */
 export type WidgetDataUpdate = Partial<ReadonlyWidgetData>;
+
+/** Widget error details */
+export type WidgetError = {
+  /** Message to display of the error */
+  message: string;
+
+  /** Type of the error */
+  type?: string;
+
+  /** Stack trace of the error */
+  stack?: string;
+
+  /** Specific error code */
+  code?: number;
+};
+
+/** Message containing a new document update */
+export const METHOD_DOCUMENT_UPDATED = 'documentUpdated';
+
+/** Message containing a document error */
+export const METHOD_DOCUMENT_ERROR = 'documentError';